Add conversion functions for WASI file flags and improve file statistics handling
- Introduced `into_wasi_fdflags` function to convert XilaFileSystemStatus to __wasi_fdflags_t.
- Updated `into_wasi_file_statistics` to maintain existing functionality while improving code structure.
- Adjusted function signatures for clarity and consistency.

+++ modules/abi/src/file_system/types.rs
@@ -59,29 +59,29 @@ impl From<file_system::Kind> for XilaFileKind {
 
 pub type XilaFileSystemMode = u8;
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_MODE_READ_MASK: u8 = file_system::Mode::READ_BIT;
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_MODE_WRITE_MASK: u8 = file_system::Mode::WRITE_BIT;
 
 pub type XilaFileSystemOpen = u8;
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_OPEN_CREATE_MASK: u8 = file_system::Open::CREATE_MASK;
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_OPEN_CREATE_ONLY_MASK: u8 = file_system::Open::EXCLUSIVE_MASK;
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_OPEN_TRUNCATE_MASK: u8 = file_system::Open::TRUNCATE_MASK;
 
 pub type XilaFileSystemStatus = u8;
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_STATUS_APPEND_MASK: u8 = file_system::Status::APPEND_BIT;
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_STATUS_NON_BLOCKING_MASK: u8 = file_system::Status::NON_BLOCKING_BIT;
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_STATUS_SYNCHRONOUS_MASK: u8 = file_system::Status::SYNCHRONOUS_BIT;
-#[no_mangle]
+#[unsafe(no_mangle)]
 pub static XILA_FILE_SYSTEM_STATUS_SYNCHRONOUS_DATA_ONLY_MASK: u8 =
     file_system::Status::SYNCHRONOUS_DATA_ONLY_BIT;
 
@@ -109,7 +109,7 @@ pub struct XilaFileSystemStatistics {
 impl XilaFileSystemStatistics {
     pub fn from_statistics(statistics: file_system::Statistics_type) -> Self {
         Self {
-            file_system: statistics.get_file_system().as_inner(),
+            file_system: statistics.get_file_system().as_inner() as u32,
             inode: statistics.get_inode().as_u64(),
             links: statistics.get_links(),
             size: statistics.get_size().as_u64(),
